refactor: organize commands
