+++ Taskfile.yml
@@ -3,11 +3,11 @@ version: '3'
 tasks:
   run:
     cmds:
-      - go run .
+      - go run ./cmd/st
 
   build:
     cmds:
-      - go build -o bin/st .
+      - go build -o bin/st ./cmd/st
 
   test:
     cmds:
