--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/fehlhabers/st/cmd"
-
-func main() {
-	cmd.Execute()
-}
